==> datasets/Ovarian_Cancer_microarray/02_prepare.ipynb <==
library(tidyverse)
source("../../utils/plots_eda.R")

# Load data

datasets <- c("GSE6008", "GSE14407", "GSE26712",  "GSE40595", "GSE36668", "GSE69428")

all_expression <- NULL
all_metadata <- NULL

for(dataset in datasets){
    metadata <- read.table(paste0("before/", dataset, "/1_metadata_short.tsv"), header = TRUE, sep = "\t")
    expr_data <- read.table(paste0("before/", dataset, "/1_expr_GBacc_MaxVar.tsv"), header = TRUE, sep = "\t", row.names = 1)

    print(paste0("Processing dataset: ", dataset, "; platform: ", metadata$platform_id[1]))
    print(paste0("Samples: ", nrow(metadata), "; Features: ", nrow(expr_data)))

    metadata <- metadata %>% mutate(Dataset = dataset) %>%
        select(Sample_geo_ID, Status, HistSubtypes, Stage2, Grade2, ClinicopathologicSubtypes, Dataset) %>%
        # outliers
        filter(!(Sample_geo_ID %in% c("GSM1701038", "GSM898305", "GSM997639"))) %>%
        rename("sample_id" = "Sample_geo_ID")

    expr_data <- expr_data[, metadata$sample_id]
    print(paste0("Samples: ", nrow(metadata), "; Features: ", nrow(expr_data)))

    #########################################################################################################################
    #  plot
    plot_res <- plot_diagnostic(expr_data, metadata, dataset,
                                log_transform = TRUE, with_rowname = TRUE)
    layout <- (plot_res[[1]] + plot_res[[2]] ) / 
              (plot_res[[3]] )
    ggsave(paste0("before/", dataset, "/diagnostic_plot.png"), 
                plot = layout, width = 12, height = 12)


    expr_data <- rownames_to_column(as.data.frame(expr_data), "row_ID")
    # save data
    if(is.null(all_metadata)){
        all_metadata <- metadata
        all_expression <- expr_data
    } else {        
        all_metadata <- rbind(all_metadata, metadata)
        all_expression <- full_join(all_expression, expr_data, by = "row_ID")
    }
    print(paste0("Combined Samples: ", nrow(all_metadata), "; Features: ", nrow(all_expression)))
    print(" ")
}

rownames(all_expression) <- all_expression$row_ID
all_expression <- all_expression[, -1]

# # plot the combined data
print("Plotting combined data")
plot_res <- plot_diagnostic(all_expression, all_metadata, "Combined",
                                log_transform = TRUE, with_rowname = TRUE)
layout <- (plot_res[[1]] + plot_res[[2]] ) / 
          (plot_res[[3]] )
ggsave("before/diagnostic_plot.png", 
            plot = layout, width = 12, height = 12)

# Save data for correction and after correction

all_metadata$batch <- as.numeric(as.factor(all_metadata$Dataset)) - 1
# all_metadata$batch <- 0
all_metadata$Status <- as.numeric(as.factor(all_metadata$Status))
all_metadata$Status = all_metadata$Status - 1


for (dataset in unique(all_metadata$Dataset)) {
    print(paste0("Save data prior to batch correction for ", dataset))
    dataset_metadata <- all_metadata[all_metadata$Dataset == dataset,]
    dataset_metadata <- dataset_metadata %>% select(sample_id, Status, batch)
    
    dataset_expression <- all_expression[, dataset_metadata$sample_id]
    dataset_expression <- na.omit(dataset_expression)
    dataset_expression <- dataset_expression[apply(dataset_expression, 1, var) > 0, ]

    dataset_expression <- dataset_expression %>% rownames_to_column("gene_id")
    print(paste0("Samples: ", nrow(dataset_metadata), "; Features: ", nrow(dataset_expression)))
    
    write.table(dataset_metadata, 
        file = paste0("before/", dataset, "/design.tsv"), sep = "\t", quote = FALSE, row.names = FALSE)
    write.table(dataset_expression, 
        file = paste0("before/", dataset, "/expr_for_correction.tsv"), sep = "\t", quote = FALSE, row.names = FALSE)

}

all_expression <- na.omit(all_expression)
all_expression <- all_expression[apply(all_expression, 1, var) > 0, ]
all_expression <- all_expression[, all_metadata$sample_id]

write.table(all_metadata %>% select(sample_id, Status, batch), 
    file = "before/all_design.tsv", sep = "\t", quote = FALSE, row.names = FALSE)
write.table(all_expression %>% rownames_to_column("gene_id"),
    file = "before/all_expr_for_correction.tsv", sep = "\t", quote = FALSE, row.names = FALSE)

for (dataset in unique(all_metadata$Dataset)) {

    dataset_metadata <- all_metadata[all_metadata$Dataset == dataset,]
    print(paste0("Save data prior to batch correction for ", dataset))
    dataset_expression <- all_expression[, dataset_metadata$sample_id]
    dataset_expression <- dataset_expression %>% t() %>% as.data.frame() %>% 
    rownames_to_column("sample_id")
    print(paste0("Samples: ", nrow(dataset_expression), "; Features: ", ncol(dataset_expression)))
    
    write.table(dataset_expression, 
        file = paste0("before/", dataset, "/kmeans/expr_for_kmeans_before.tsv"), sep = "\t", quote = FALSE, row.names = FALSE)

}
