==> datasets/Ovarian_Cancer_microarray/03_correction.ipynb <==
library(tidyverse)
library(sva)

source("../../utils/plots_eda.R")

# Load data

datasets <- c("GSE6008", "GSE14407", "GSE26712",  "GSE40595", "GSE36668", "GSE69428")

all_metadata <- read.table("before/all_design.tsv", header = TRUE, sep = "\t")
all_expression <- read.table("before/all_expr_for_correction.tsv", header = TRUE, sep = "\t") %>%
    column_to_rownames("gene_id")

all_expression <- all_expression[, all_metadata$sample_id]

# remove NAs rows
all_expression <- na.omit(all_expression)
# remove genes with 0 variance
all_expression <- all_expression[apply(all_expression, 1, var) > 0, ]

print("Loading metadata and expression data")
print("Metadata dimensions:")
print(dim(all_metadata))
print("Expression data dimensions:")
print(dim(all_expression))

design <- model.matrix(~all_metadata$Status)

corrected_expr <- sva::ComBat(dat = all_expression, 
                              batch = all_metadata$batch, 
                              mod = design)

corrected_expr <- as.data.frame(corrected_expr)


all_metadata$Status <- all_metadata$Status
all_metadata$Dataset <- as.character(all_metadata$batch)


# plot the combined corrected data
print("Plotting combined corrected data")
plot_res <- plot_diagnostic(corrected_expr, all_metadata, "Combined Corrected",
                            log_transform = TRUE, with_rowname = TRUE)
layout <- (plot_res[[1]] + plot_res[[2]] ) / 
          (plot_res[[3]] )
ggsave("after/diagnostic_plot_corrected.png", 
            plot = layout, width = 12, height = 12)


# write out the corrected expression data
write.table(corrected_expr %>% rownames_to_column("gene_id"), "after/all_corrected_R_expr.tsv", sep = "\t", 
            quote = FALSE, col.names = TRUE, row.names = FALSE)
